Revert "CAS-603 Implement timeline for probation practitioners"
