+++ server/paths/api.ts
@@ -67,8 +67,6 @@ const oasysPath = personPath.path('oasys')
 
 const cas3Path = path('/cas3')
 const reportsCas3Path = cas3Path.path('reports')
-const timelineCas3Path = cas3Path.path('timeline').path(':assessmentId')
-
 const applyPaths = {
   applications: {
     show: singleApplicationPath,
@@ -130,7 +128,6 @@ export default {
       create: clarificationNotePaths.notes,
       update: clarificationNotePaths.notes.path(':clarificationNoteId'),
     },
-    timeline: timelineCas3Path,
   },
   people: {
     risks: {
